Use POSIX semaphores and POSIX shared memory for fastdb
Switched the fastdb to use POSIX IPC instead of SysV IPC.

The POSIX named semaphores show up under /dev/shm.
I wanted to name the semaphores based on the path
of the imdb directory, but sem_open() doesn't allow
slashes in the semaphore name.  So instead of the
full path, the naming convention I used was:

/dev/shm/sem.{ftok}-imdb.{type}

Where {ftok} is the value returned from ftok(path, 0)
with the path set to the directory that holds the imdb.dm
(database monitor, aka the shared memory) and the imdb.odb
(the data itself).

The semaphore types are:
  .in Initialization
  .rs Read 
  .us Update
  .ws Write
  .bce Database Events

Fixed issues in the fastdb's POSIX implementation,
including handling EINTR and calling sem_post()
the correct number of times,
and a nasty situation that existed if the Initialization
semaphore previously existed, but the shared memory
didn't.

Made sure the POSIX implementations use the keyFileDir
path to create the imdb.dm and imdb.odb files.

Because of the {ftok} part, tests which use their own
database (by setting the context directory, via keyFileDir) 
should not end up conflicting with other tests or the "real"
database.

Changed system startup to delete /dev/shm/sem.*-imdb.*
on startup, along with the /var/sipxdata/tmp/imdb.* files.
This will delete ALL imdb semaphores (even test ones), 
so starting the system up while running tests is not a
good idea!

There isn't a need for ipcclean.pl anymore, but it doesn't
hurt to keep it for now.

The switch to POSIX is being done as an experiment to see
if the issues we are seeing with the fastdb (XECS-619) 
are possibly SysV IPC related.

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@11411 ab1d8caa-1f67-47f1-9e81-24633a41865c