Add Group Page Timeout configuration
Patch from Laurentiu Ceausescu.  Thanks.

Changed the patch around a bit.  Made the timeout an int inside sipXpage
instead of a string.  Made the timeout value "0" mean no timeout.

In sipXconfig, I changed a new page group to default to enabled and a with
a 60 second timeout.  I also changed the values of page number to allow 0 
(it is legal).

Fixes XECS-1260 Group Page server page interval should be configurable

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@12537 ab1d8caa-1f67-47f1-9e81-24633a41865c