XCF-2648: Made new conference persistent so that changes aren't lost when leaving the page to assign an owner.
The conference property of EditConference is now persistent in the session, but this doesn't last very long. The pageBeginRender() method will clear this persistent conference unless we are coming back from selecting a user. The page knows this by checking for the presence of a value for getSelectedUsers() and clearing it if there is none.


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@13199 ab1d8caa-1f67-47f1-9e81-24633a41865c