XCF-2797: Implemented a better internal representation for model features.
This adds a Set<String> denoting particular model-specific features to a DeviceDescriptor.
The bean entry in a *-beans.xml for a model can now denote its supported features as a <set>:

<property name="supportedFeatures">
  <set>
    <value>disableCallList</value>
    <value>voiceQualityMonitoring</value>
  </set>
</property>

Each supported feature gets added as a "definition" to use for conditional settings, i.e.:
<group name="qualityMonitoring" if="voiceQualityMonitoring">
 ...
</group>

Patch submitted by Bogdan Brezoi - thanks!

(I made a slight change to Bogdan's patch - moved the supportedFeatures property into the
common base class DeviceDescriptor so that it doesn't have to be repeated in the different
model classes.)


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@13649 ab1d8caa-1f67-47f1-9e81-24633a41865c