XCF-2698 LDAP Imported data does not the preview if mapping has been changed
The m_userMapper.m_attrMap in class LdapRowInserter is not explicitly set up. It is set up once only by:

LdapRowInserter.insertRow(sr, Attributes)
    UserMapper.getUserName(attt)
        UserMapper.getAttrMap() {
            if (m_attrMap != null) {
                 return m_attrMap;
            } else
                m_attrMap = m_ldapManager.getAttrMap() /* set up by this code. */
        }

So the attributes will always be the same as the first LDAP import carried out, leading to the
problem, where the preview looks OK, but the actual Import doesn't work as expected.

The fix is to set LdapRowInserter.m_userMapper.m_attrMap in LdapRowInserter.beforeInserting() to the
m_ldapManager.m_attrMap, which has itself been set up when the attributes have been selected.

We could use the m_attrMap of LdapRowInserter (as this has been set up to the correct value in
LdapImportManagerImpl.insert()) to initialise the LdapRowInserter.m_userMapper.m_attrMap if desired.

Patch from Paul McDaid - thanks!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@13884 ab1d8caa-1f67-47f1-9e81-24633a41865c