Relax autoattendants.xsd schema to match sipXconfig default.
Allowes for "empty" <holiday>, <regularhours>, and <afterhours> elements.

The "id" subelement is required, but only if there are other subelements.  So
set minOccurs to be 0.

The "default" autoattendant on newly installed systems has empty <holiday>
and <regularhours> elements, and as such was failing configuration tests
as this didn't match the schema.  As the empy configuration is valid, relax
the schema to reflect this.

Fixes XECS-1833 - Auto Attendant 100 access fails after a split second

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@13996 ab1d8caa-1f67-47f1-9e81-24633a41865c