Rework validusers.xml to reflect modern design.
This version is cleaner and shorter.  Instead of duplicating the <user> for
each alias (which was an emulation of the imdb table the old AA used),
this one creates a list of <aliases> under <user>.

Generator of validusers.xml (sipXconfig) and consumer (sipXivr) both 
modified to match. validusers.xsd.in updated accordingly.

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@14008 ab1d8caa-1f67-47f1-9e81-24633a41865c