XCF-2945, XCF-2946, XCF-2949: First iteration of software package update functionality.
Currently, the download/install process is performed synchronously after
stopping the server, so updating the packages can result in a fair
amount of server downtime.

This will be resolved in the near future as the download process, which
takes the bulk of the time, will be made asynchronous and only the
package installation (which is fairly quick) will be done synchronously
after stopping the server, to minimize downtime.

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@14043 ab1d8caa-1f67-47f1-9e81-24633a41865c