XCF-3027 change UI to push configurations to the Server
Servers are like phones: generating configuration file for the Server now resembles generating
cnfiguration file for a phone. There is a "Send Profile" button on a server page.

The implementation of generating configuration files for a server is still flawed: instead of
sending files to a selected servers only we are generating configuration for all services that are
running on that server and pushing this configuration to all servers.


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@14085 ab1d8caa-1f67-47f1-9e81-24633a41865c