XCF-3053i simple REST service for placing calls
The URL looks like this:

https://{host}:8443/sipxconfig/rest/call/{number}

HTTP PUT to this URL will cause sipXconfig to place call to {number}.
The calls is placed using authorized user credentials. It works in the
same way as click-to-call available on the user portal. User's phone
rings first and, once someone answers, it places the call to {number}.

To test the service with curl use:
curl -k -X PUT https://{user}:{password}@{host}:8443/sipxconfig/rest/call/{number}

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@14706 ab1d8caa-1f67-47f1-9e81-24633a41865c