sipXconfig: only use select AliasProviders when generating ValidUsers.xml
An explicit list had been passed in, but the AliasCollector was still
including auto-discovered AliasProviders even though it was supposed
to ignore them.  Thus additional Alias entries were making it into 
ValidUsers.xml, and the AA was picking the wrong one at times.

Fixes XECS-2310-Auto attendant transfer does not work as expected when the call is being forwarded

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@14878 ab1d8caa-1f67-47f1-9e81-24633a41865c