XX-5942 incremental phonebook search
-GWT suggest box calls rest service
-Integrate new GWT component in user phonebook page
-Localization example
-declare .css file in .xml GWT module instead of Tapestry phonebook page

Limitations: need to force user/password to be sent in http header when
rest service is called Currently browser requests the first time to log
in. Next times it remembers username/password and automatically resends
them.
The interaction between the search box and the table is not ideal yet,
but we are rewriting the table as GwT component as well.

Patch from: Mircea Carasel - thanks!

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@16154 ab1d8caa-1f67-47f1-9e81-24633a41865c