sipXivr: correct cancel behavior when entering an extension
Fixes (parts of):
   XX-6541 - Incorrect prompt is played when * key is pressed to cancel 
   the recorded greeting .

The parts in XX-6541 that deal with canceling during "dialing an extension" 
are fixed.

The first part in XX-6541 that deals with cancel during recorded
greeting is not changed...the cancel concept always goes up one level, 
not two.  The behvior of pressing cancel after recording a greeting 
goes back to the recording choice, not all the way back to the Voicemail
Options menu.  It took two choices to get in, it takes two cancels
to get out.

(2) Press option 5 for VM option. 
(3) Press 1 to record a personal greeting. 
(4)    Press 1 to record a standard greeting. 
(5)       Record the greeting and Press #. 
(6)       Plays the prompt "......To listen to this greeting press 1,to use this greeting press2 ,to delete this greeting press 3 ,to cancel press *.... " 
(7)       Press *

The * pressed in (7) jumps back to (4).  Press * at (4), it jumps back to (3).
Press * at (3), it jumps back to (2) the main menu.

--Woof!



git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@16573 ab1d8caa-1f67-47f1-9e81-24633a41865c