XX-6476 improve detection of proxyHost in SipStackBean
When sending SIP messages sipXconfig assumed that proxy runs on a
primary server: it now finds the server on which proxy runs by
consulting LocationsManager.

It resolves problems with setting up "click-to-call" connections and
resetting phones when Primary SIP Router role is enabled on a
non-primary server.

Based on a patch from: Constantin Tambrea - thanks!

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@16900 ab1d8caa-1f67-47f1-9e81-24633a41865c