XX-7108 Fix exceptions on the alarm history page
Move AlarmHistory table outside of the Server/Date form. If table is in
the form Tapestry expects table model to be avaialble and in some cases
it accesses the model before "pageBeginRender" for our component is
called.

Implement standard getCached/setCached idiom to delay loading alaram
events.

Remove HistoryAlarmsTableModel and use syntetic model constructed by
Tapestry from our alarm events list. We were not using the model for
anything and it was interfeering with table sorting. (The initial idea
was to use the model to cache events list but that was never
implemented).

Current implementation is still far from optimal, we retrieve the events
from the server once per render cycle instead of caching the results and
retrieve the only when the get 'old'.

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@17439 ab1d8caa-1f67-47f1-9e81-24633a41865c