XX-6777: Alarm settings (Group & Min Threshold) lost upon sipXsupervisor RPM re-install
The problem was that the Alarm Group and Threshold values were being stored on in
sipXalarms-config.xml (the replicated file), which is over-written upon a sipXsupervisor
install.

- Now storing the Alarm Group and Threshold values in a new alarm_code DB table.
- Upon loading the Diagnostics -> Alarms -> Alarm types table, the sipXalarms-config.xml
  content is merged with the alarm_code DB table.

One minor limitation still exists.  A re-install of sipXsupervisor will still always
over-write the existing sipXalarms-config.xml with a default version.  The configured
values will be restored (from the DB) upon the next replication of this file.  But (and
here's the limitation), unless it is a formal upgrade situation, the next replication
will not be done automatically.  You need to trigger it manually: An Apply on the Alarm
types table, or sipxconfig.sh --first-run (and sipXconfig restart) will do the trick.

(This isn't a problem in formal upgrade, because there's always a sipxconfig.sh
--first-run, which causes replication.)

Patch from David Saint, thanks!

git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@17480 ab1d8caa-1f67-47f1-9e81-24633a41865c