XX-8294 Phonebook : No provision to download additional contacts in admin portal
Added a component that provides a link to download the phonebook in vCard format
Since phonebook entries are now kept in the DB no matter their method of inserting in the phonebook,
it is no longer possible to download the uploaded vcf or csv files.

I have added a text which explains the new behavior on the edit phonebook page.

*note that due to an unknown bug, if an exception is thrown at manager level the error is not displayed to the user


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/branches/4.2@18807 ab1d8caa-1f67-47f1-9e81-24633a41865c