[PATCH] XX-9431 - aliases validation problems for paging groups, Registry and Presence services - added alias validation for the services - added alias validation for paging group prefix number and paging group number - alias for paging groups is formed by prefix and p.g. number - i.e. if we have a user 201 we cannot have an alias prefix 20 and a p.g. with number 1
Note that this patch made some apparently non-related changes: -
SipxHibernateDaoSupport.saveBeanWithSettings(BeanWithSettings) by introducing save and merge instead of saveOrUpdate. This change was necessary due to org.hibernate.NonUniqueObjectException: a different object with the same identifier value was already associated with the session thrown when trying to save Presence or Registry changes.

Introduced a new method in SipxHibernateDaoSupport.saveOrUpdateBeanWithSettings that makes use of Hibernate's saveOrUpdate. This was necessary due to a problem saving GeneralPhonebookSettings