XX-9663 Improvement of resource-lists.xml generation time for large number of users
ResourceLists.java loads all users in order to generate resource-lists.xml
It does so by useing forAllUsersDo() method which load users from PGSQL db.

With this patch users are loaded from mongo. This will significantly improve the time it takes to generate the file.

- SpeedDials data set was added to add the speed dial info to mongo
- ValidUsers method getUsers has been refactored to retrieve all users from mongo - old method was named getValidUsers()
- tests have been updated/added

- some unneeded xml test files were removed

NOte that with introduction of the new data set, the time needed to regenerate the entier mongo db has increased.