XX-9689 - import slowing down after a number of users.
turned out that the filling of Hibernate session is the culprit for this
Clearing the session would improve timings significantly.

** note that we need to issue a session flush before clear for the users to be inserted into the DB.
Clearing the session would "cancel all pending saves, updates and deletions". It is actually a race
between auto-flush and clear, but I haven't seen flush win, so we need to make sure data is inserted
in DB.