XX-9889: OF plugin: @conf is sending the invitation twice - your phone rings twice
openire loads DefaultMessageInterceptor and then looks in extras path to search for other interceptors and to load them.
but extras contains symbolic link that points to the of presence plugin jar

-removed symbolic link from /opt/openfire/extras/sipxecs/ - that was pointing to the sipxecs "share" plugins directory
where of sipxpresence.jar is located, since there is another symbolic link in openfire/plugins directory that points to the
of presence jar. the openfire/plugins is the standard path from where openfire picks /uarchives and loads plugins.
The extras/lib can remain and can be used for other libraries that we want to be loaded. If those libraries
will contain message interceptors they will automatically loaded in openfire

of presence plugin looks in extras only when openfire.home varable is set.