beware of cfengines "delete_lines", it does not do what you think it does
If you do this :

bundle edit_line foo {
  delete_lines:
    "a";
    "b";
}

does not delete any lines that have "a" or "b", it delete's lines that
have "a" or "b" if file only has *both* "a" and "b" in the file and in
order of "a" then "b".  In summary, it's almost useless to have
multiple lines.

What you probably want to do is this:

bundle edit_line foo {
  delete_lines:
    "(a|b)";
}

Also, remember, it's "anchored regex", so if any line the has an "a"
anywhere in the line, then it's

bundle edit_line foo {
  delete_lines:
    "^.*(a|b).*$";
}