XX-10459 Park service crash if Max Session is set to 6 or below it
Adjusted the value of MaxSession to fix park server crashes.
1. sipXpark uses sipxMediaLib which will assert if the given number of audio buffers is below 34.
Audio buffers number is computed as 6 * MaxSessions. So for MaxSession between [1,6] to avoid the
assert failure the number of audio buffers provided will be 42 (6 * 7).

2. MaxSession value should be only positive numbers as it makes no sense to have negatives or 0.
So if the configured value is <=0 the default value 50 will be used.

XX-10492 was opened to clarify the usage off MaxSession values in the Park Server configuration
interface.