PT42039311 - Modify SipSubscriptionClient class to use boost::shared_ptr and standard containers to get rid of locks.
All containers from this class were migrated to std::vector and std::map. The pointers stored in this containser were wrapped
in a boost::shared_ptr which ensure safe deletion. Locks from this class were reorganized. Fewer locks are needed now because
of shared_ptr: 1. lock access to the containers; 2. lock a stored pointer while it is retrieved from container and used.
Overall class speed improved because locks are now only per stored pointer and not per class method (as they were before).

A wrapper class was implemented to lock a pointer after it was retrieved from container and ensure that only one thread is using it.
The wrapper is non-copyable (because of the inner locks) so it is enveloped in a std::auto_ptr for easy passing of the wrapper object.

Some functions were refactored.
Added a missing header in the unit test to make test compile.