PT39527565 - If mongo is unavailable, SAA and RLS service should send 500 error instead of shutting down
Added checks for mongo connectivity in both RLS and SAA. If mongo is not available upon
start the application will log an error and will exit. After that the app will rely on
the supervisor to restart it so that it can re-check mongo connectivity.

This solution was preferred instead of checking mongo connectivity in a loop because
of mongo C++ driver issues.