PT47131989 - Zendesk #367 SAA stops working
The issues appeared when subscriptions were refreshed, that is once per day. After refresh
the subscribe mappings were deleted so SAA could not remap the new subscriptions to proper
handles. This was fixed and now the subscribe mapping is not deleted anymore so the reg
subscription can find the proper handle after it is refreshed.

Termination of SAA appearances to phones was improved so that the appearance handler is
deleted only after the parent subscription was ended - this also cleared out some error messages.

Changed the log level to ERR for some log messages.