PT46534249 - Callpark - The call is not connecting to call park extension
Collateral bug because of commit ad32e34867ef0e96232324a7a885062920948800.
Because the initialization of sipXpark was split in two code blocks some
variables which were allocated on the stack in the first block were going
out of scope in subsequent code blocks making the functionality unstable.

This broken behavior was fixed by rejoining the code blocks in one large block.