PT48083963 - Calls hold/pickup fails in a HA setup with two servers and shared lines
Two changes were done in proxy's NAT traversal plugin:
- Use a modified Contact header (the header was modified to contain endpoint's public
and private address) for all message types except for SUBSCRIBE and NOTIFIES. Previously,
the header was restored for all messages except INVITES in order to fix some presence
issues. Because of this approach a proxy was unable to route correctly other requests
such as PRACK or ACK.
- Do not let a proxy change the request URI of a request if it is not the "last" proxy
to route that request to the endpoint behind NAT. If a proxy changes the request URI
the next proxy will not be able to route it correctly.