PT50360093 - Core dump in sipxsaa on openuc.ezuce.com
sipXsaa crashed because an object was deleted while in usage. After deletion
the object was accessed again leading to crash (A refresh event was triggered
by the stack for an Appearance subscription, which triggered a Appearance::subscriptionEventCallback
with subscription state set to terminated. This callback wrongfully set the
_isTerminated flag to true. Because of this the next callback Appearance::subscriptionEventCallback
with subscription state set to terminated  would lead to deletion of
the Appearance and a crash.)

To fix the problem appearances should be deleted only when these conditions are met:
1. terminate() was requested by parent reg subscription (from AppearanceGroup);
2. Appearance::subscriptionEventCallback with subscription state set to
terminated was triggered after step 1.

_isTerminated flag was replaced with a termination state condition.