PT51082409 - Lots of core dumps with mongo crashes on openuc.ezuce.com
Added catch_global in RLS, SAA, Publisher so mongo exception will be catched
and logged before aborting execution.
Added try catch for mongo exceptions in SAA message handling loop to respond
with 500 to messages for existing subscriptions or notifications.
Publisher will check at startup if mongo is alive, if not it will exit.