UC-690: HTML5 tags/checks in voicemail playback
-extended config voicemail api to show file format
-fixed mime type given the audio file format (audio/mpeg for mp3 or audio/wav for wav) it was audio/x-wav always. IE and probably other browsers does not recognize x-wav or x-mpeg
-show internal/external player option given browser support
-extended browser acceptance to IE 10, Opera when login

Tested with IE 8, 9, Firefox 18, 22, 23, Safari 5, 6, Opera 12, 15 on Linux, windows, mac, android platforms - works fine. The only drawback here is that some versions of Opera 12 on Linux have trouble playing .wav with HTML 5, despite the fact that they mention that Opera 10+ supports wav playing. I also noticed that with Safari

when a browser does not support playing and user-portal is not able to detect this, it will automatically fallback to external player option