UC-1185 Investigate log entry "received response without transaction"
Lowered the log level from WARNING to INFO. This message is logged for
every response passing the UA for which a live transaction could not be found.
We chose to keep it on INFO level and not lower it to  DEBUG level
because this log is a way of measuring the proxy's effort of forwarding
responses with no transaction. However on some scenarios it should be expected
because proxy's NatMaintainer does keep alive by sending an OPTION request with
no transaction, so, for every OPTION message send by NatMaintainer this
log message will be logged.