UC-4436: Decrease default number of backups to retain
set backup limit default to 5.
when system is first installed, it will use value 5 as default
also when you run sipxecs-setup --reset-all same value 5 will be used no matter
what the previous was.
when you upgrade the system (yum update) the existing value will be used, it will not revert to 5